requirements.txt
# Django core
Django==5.1.6
django-environ==0.12.0
dj-database-url==2.2.0

# REST & authentication
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
bcrypt==4.2.0

# UI & tweaks
django-widget-tweaks==1.4.8
django-cors-headers==4.4.0

# Databases
mysqlclient==2.2.4       # For local MySQL
psycopg2-binary==2.9.9    # For Neon PostgreSQL

# Redis & background tasks
django-redis==5.4.0
redis==5.1.1

# Media storage
cloudinary==1.44.0
django-cloudinary-storage==0.3.0

# WebSockets (if needed)
channels==4.2.2

# Networking & APIs
requests==2.32.3
dj-database-url==2.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
celery==5.3.6

python-decouple==3.8
